Moved param decoder into separate module.
