+++ server/routes/results.js
@@ -3,7 +3,7 @@ const {
   getAreaInHectares,
   getCentreOfPolygon
 } = require('../services/shape-utils')
-const { decode } = require('@mapbox/polyline')
+const { decodePolygon } = require('../services/param-polygon-decoder')
 
 module.exports = [
   {
@@ -13,30 +13,20 @@ module.exports = [
       description: 'Results Page',
       handler: async (request, h) => {
         const { polygon } = request.query
-        let parsedPolygon
-        try {
-          const confirmedJSON = JSON.parse(polygon)
-          if (Array.isArray(confirmedJSON) && Array.isArray(confirmedJSON[0])) {
-            parsedPolygon = polygon
-          } else {
-            throw new Error('not a polygon array')
-          }
-        } catch {
-          const decodedResult = decode(polygon)
-          parsedPolygon = JSON.stringify(decodedResult)
-        }
+        const decodedPolygon = decodePolygon(polygon)
+        
         const [contactData, floodData] = await Promise.all([
-          request.server.methods.getPsoContactsByPolygon(parsedPolygon),
-          request.server.methods.getFloodDataByPolygon(parsedPolygon)]
+          request.server.methods.getPsoContactsByPolygon(decodedPolygon),
+          request.server.methods.getFloodDataByPolygon(decodedPolygon)]
         )
         const showOrderProduct4Button = config.appType === 'internal' || contactData.useAutomatedService === true
-        floodData.areaInHectares = getAreaInHectares(parsedPolygon)
-        floodData.centreOfPolygon = getCentreOfPolygon(parsedPolygon)
+        floodData.areaInHectares = getAreaInHectares(decodedPolygon)
+        floodData.centreOfPolygon = getCentreOfPolygon(decodedPolygon)
         floodData.isFZ1Andlt1ha = floodData.floodZone === '1' && floodData.areaInHectares < 1
         floodData.isFZ1Andgt1ha = floodData.floodZone === '1' && floodData.areaInHectares >= 1
         floodData.areaInHectares = floodData.areaInHectares !== '0' && floodData.areaInHectares !== 0 ? floodData.areaInHectares : 'less than 0.01'
         floodData.riversAndSea = floodData.floodZone !== '1' || floodData.floodZoneClimateChange || floodData.floodZoneClimateChangeNoData
-        return h.view('results', { parsedPolygon, floodData, contactData, showOrderProduct4Button })
+        return h.view('results', { decodedPolygon, floodData, contactData, showOrderProduct4Button })
       }
     }
   }
